The test of existence of directory must be in forceDelete methods
git-svn-id: https://svn.apache.org/repos/asf/maven/components/trunk@163012 13f79535-47bb-0310-9956-ffa450edef68

+++ maven-mboot2/src/main/java/FileUtils.java
@@ -912,6 +912,11 @@ public class FileUtils
     public static void forceDelete( final File file )
         throws IOException
     {
+        if ( ! file.exists() )
+        {
+            return;
+        }
+
         if ( file.isDirectory() )
         {
             deleteDirectory( file );
@@ -934,6 +939,11 @@ public class FileUtils
     public static void forceDeleteOnExit( final File file )
         throws IOException
     {
+        if ( ! file.exists() )
+        {
+            return;
+        }
+
         if ( file.isDirectory() )
         {
             deleteDirectoryOnExit( file );
